# openapi.json
{
  "components": {
    "schemas": {
...
      "CacheMeta": {
        "description": "Whether a result was served from the local cache or fetched live.\n\n>>> CacheMeta(hit=False, source=\"live\").source\n'live'",
        "properties": {
          "hit": {
            "title": "Hit",
            "type": "boolean"
          },
          "source": {
            "enum": [
              "live",
              "cache"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "hit",
          "source"
        ],
        "title": "CacheMeta",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestState": {
        "enum": [
          "active",
          "ignored"
        ],
        "type": "string"
      },
...
      "QuotaMeta": {
        "description": "The day's quota budget snapshot a guarded call reports.\n\n>>> QuotaMeta(limit=100, used=3, remaining=97).remaining\n97",
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "remaining": {
            "title": "Remaining",
            "type": "integer"
          },
          "used": {
            "title": "Used",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "used",
          "remaining"
        ],
        "title": "QuotaMeta",
        "type": "object"
      },
...
      "TranscriptDecisionOutcomeRead": {
        "description": "The transcript status after a human decision.",
        "properties": {
          "transcript_status": {
            "$ref": "#/components/schemas/TranscriptStatus"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "transcript_status"
        ],
        "title": "TranscriptDecisionOutcomeRead",
        "type": "object"
      },
...
      "TranscriptStatus": {
        "enum": [
          "pending",
          "retrying",
          "needs_review",
          "available",
          "unavailable"
        ],
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "YouTubeSource": {
        "const": "liked",
        "type": "string"
      },
...
      "YouTubeTranscriptResponse": {
        "description": "A fetched transcript: the updated video, its text, and quota + cache.",
        "properties": {
          "cache": {
            "$ref": "#/components/schemas/CacheMeta"
          },
          "quota": {
            "$ref": "#/components/schemas/QuotaMeta"
          },
          "transcript": {
            "title": "Transcript",
            "type": "string"
          },
          "video": {
            "$ref": "#/components/schemas/YouTubeVideoRead"
          }
        },
        "required": [
          "video",
          "transcript",
          "quota",
          "cache"
        ],
        "title": "YouTubeTranscriptResponse",
        "type": "object"
      },
      "YouTubeVideoListResponse": {
        "description": "A browse/search result: the videos plus the call's quota + cache.",
        "properties": {
          "cache": {
            "$ref": "#/components/schemas/CacheMeta"
          },
          "quota": {
            "$ref": "#/components/schemas/QuotaMeta"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/YouTubeVideoRead"
            },
            "title": "Videos",
            "type": "array"
          }
        },
        "required": [
          "videos",
          "quota",
          "cache"
        ],
        "title": "YouTubeVideoListResponse",
        "type": "object"
      },
      "YouTubeVideoRead": {
        "description": "HTTP representation of an ingested video, exposing its derived state.\n\n>>> read = YouTubeVideoRead(\n...     id=\"018f0000-0000-7000-8000-000000000000\",\n...     video_id=\"v1\",\n...     source=\"liked\",\n...     state=\"active\",\n...     title=\"Talk\",\n...     channel=\"PyConf\",\n...     topic=\"python\",\n...     description=\"\",\n...     transcript=None,\n...     transcript_status=\"pending\",\n...     created_at=datetime(2026, 1, 1),\n...     updated_at=datetime(2026, 1, 1),\n...     ignored_at=None,\n... )\n>>> read.state\n'active'",
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid7",
            "title": "Id",
            "type": "string"
          },
          "ignored_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ignored At"
          },
          "source": {
            "$ref": "#/components/schemas/YouTubeSource"
          },
          "state": {
            "$ref": "#/components/schemas/IngestState"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "transcript_status": {
            "$ref": "#/components/schemas/TranscriptStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "video_id",
          "source",
          "state",
          "title",
          "channel",
          "topic",
          "description",
          "transcript",
          "transcript_status",
          "created_at",
          "updated_at",
          "ignored_at"
        ],
        "title": "YouTubeVideoRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Tether",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/youtube": {
      "get": {
        "description": "List active ingested videos, optionally filtered by topic and source.",
        "operationId": "browse_youtube_api_youtube_get",
        "parameters": [
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/YouTubeSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeVideoListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Browse Youtube"
      }
    },
    "/api/youtube/search": {
      "get": {
        "description": "Keyword Search across saved content and transcript text.",
        "operationId": "search_youtube_api_youtube_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeVideoListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Youtube"
      }
    },
...
    "/api/youtube/{video_id}/ignore": {
      "post": {
        "description": "Purge a video from ingestion.",
        "operationId": "ignore_youtube_video_api_youtube__video_id__ignore_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeVideoRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ignore Youtube Video"
      }
    },
    "/api/youtube/{video_id}/retry": {
      "post": {
        "description": "Un-ignore a previously purged video.",
        "operationId": "retry_youtube_video_api_youtube__video_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeVideoRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Youtube Video"
      }
    },
    "/api/youtube/{video_id}/transcript": {
      "post": {
        "description": "Fetch and persist a transcript for an ingested video.",
        "operationId": "fetch_youtube_transcript_api_youtube__video_id__transcript_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeTranscriptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Youtube Transcript"
      }
    },
    "/api/youtube/{video_id}/transcript-decision/give-up": {
      "post": {
        "description": "Confirm that a review-needed video has no transcript worth pursuing.",
        "operationId": "give_up_transcript_api_youtube__video_id__transcript_decision_give_up_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptDecisionOutcomeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Give Up Transcript"
      }
    },
    "/api/youtube/{video_id}/transcript-decision/keep-trying": {
      "post": {
        "description": "Return a review-needed transcript to pending acquisition.",
        "operationId": "keep_trying_transcript_api_youtube__video_id__transcript_decision_keep_trying_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptDecisionOutcomeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Keep Trying Transcript"
      }
    }
  }
}